Add files via upload
Sixty-seven maps from “Indian Land Cessions in the United States,” compiled by Charles C. Royce and published as the second part of the two-part Eighteenth Annual Report of the Bureau of American Ethnology to the Secretary of the Smithsonian Institution, 1896-1897 have been scanned, georeferenced in JPEG2000 format, and digitized to create this feature class of cession maps. The mapped cessions and reservations included in the 67 maps correspond to entries in the Schedule of Indian Land Cessions, “indicating the number and location of each cession by or reservation for the Indian tribes from the organization of the Federal Government to and including 1894, together with descriptions of the tracts so ceded or reserved, the date of the treaty, law or executive order governing the same, the name of the tribe or tribes affected thereby, and historical data and references bearing thereon,” as set forth in the subtitle of the Schedule. Go to this URL for full metadata: http://data.fs.usda.gov/geodata/edw/edw_resources/meta/S_USA.TRIBALCEDEDLANDS.xml Each Royce map was georeferenced against one or more of the following USGS 1:2,000,000 National Atlas Feature Classes contained in \NatlAtlas_USGS.gdb: cities_2mm, hydro_ln_2mm, hydro_pl_2mm, plss_2mm, states_2mm. Cessions were digitized as a file geodatabase (GDB) polygon feature class, projected as NAD83 USA_Contiguous_Lambert_Conformal_Conic, which is the same projection used to georeference the maps. The feature class was later reprojected to WGS 1984 Web Mercator (auxiliary sphere) to optimize it for the Tribal Connections Map Viewer. Polygon boundaries were digitized as to not deviate from the drawn polygon edge to the extent that space could be seen between the digitized polygon and the mapped polygon at a viewable scale. Topology was maintained between coincident edges of adjacent polygons. The cession map number assigned by Royce was entered into the feature class as a field attribute. The Map Cession ID serves as the link referencing relationship classes and joining additional attribute information to 752 polygon features, to include the following: 1. Data transcribed from Royce’s Schedule of Indian Land Cessions: a. Date(s), in the case of treaties, the date the treaty was signed, not the date of the proclamation; b. Tribe(s), the tribal name(s) used in the treaty and/or the Schedule; and c. Map Name(s), the name of the map(s) on which a cession number appears; 2. URLs for the corresponding entry in the Schedule of Indian Land Cessions (Internet Archive) for each unique combination of a Date and reference to a Map Cession ID (historical references in the Schedule are included); 3. URLs for the corresponding treaty text, including the treaties catalogued by Charles J. Kappler in Indian Affairs: Laws and Treaties (HathiTrust Digital Library), executive order or other federal statute (Library of Congress and University of Georgia) identified in each entry with a reference to a Map Cession ID or IDs; 4. URLs for the image of the Royce map(s) (Library of Congress) on which a given cession number appears; 5. The name(s) of the Indian tribe or tribes related to each mapped cession, including the name as it appeared in the Schedule or the corresponding primary text, as well as the name of the present-day Indian tribe or tribes; and 6. The present-day states and counties included wholly or partially within a Map Cession boundary. During the 2017-2018 revision of the attribute data, it was noted that 7 of the Cession Map IDs are missing spatial representation in the Feature Class. The missing data is associated with the following Cession Map IDs: 47 (Illinois 1), 65 (Tennessee and Bordering States), 128 (Georgia), 129 (Georgia), 130 (Georgia), 543 (Indian Territory 3), and 690 (Iowa 2), which will be updated in the future. This dataset revises and expands the dataset published in 2015 by the U.S. Forest Service and made available through the Tribal Connections viewer, the Forest Service Geodata Clearinghouse, and Data.gov. The 2018 dataset is a result of collaboration between the Department of Agriculture, U.S. Forest Service, Office of Tribal Relations (OTR); the Department of the Interior, National Park Service, National NAGPRA Program; the U.S. Environmental Protection Agency, Office of International and Tribal Affairs, American Indian Environmental Office; and Dr. Claudio Saunt of the University of Georgia. The Forest Service and Dr. Saunt independently digitized and georeferenced the Royce cession maps and developed online map viewers to display Native American land cessions and reservations. Dr. Saunt subsequently undertook additional research to link Schedule entries, treaty texts, federal statutes and executive orders to cession and reservation polygons, which he agreed to share with the U.S. Forest Service. OTR revised the data, linking the Schedule entries, treaty texts, federal statues and executive orders to all 1,172 entries in the attribute table. The 2018 dataset has incorporated data made available by the National NAGPRA Program, specifically the Indian tribe or tribes related to each mapped cession, including the name as it appeared in the Schedule or the corresponding primary text and the name of the present-day Indian tribe or tribes, as well as the present-day states and counties included wholly or partially within a Map Cession boundary. This data replaces in its entirety the National NAGPRA data included in the dataset published in 2015. The 2015 dataset incorporated data presented in state tables compiled from the Schedule of Indian Land Cessions by the National NAGPRA Program. In recent years the National NAGPRA Program has been working to ensure the accuracy of this data, including the reevaluation of the present-day Indian tribes and the provision of references for their determinations. Changes made by the OTR have not been reviewed or approved by the National NAGPRA Program. The Forest Service will continue to collaborate with other federal agencies and work to improve the accuracy of the data included in this dataset. Errors identified since the dataset was published in 2015 have been corrected, and we request that you notify us of any additional errors we may have missed or that have been introduced. Please contact Rebecca Hill, Policy Analyst, U.S. Forest Service, Office of Tribal Relations, at rebeccahill@fs.fed.us with any questions or concerns with regard to the data included in this dataset.